requirements.txt
djangorestframework==3.6.2
Django==1.9.7
Pillow==3.3.1
pyjwt==1.4.2
MySQL-python==1.2.5
djangorestframework-jwt==1.4.2